+++ tests/Endpoint/AccountEndpointTest.php
@@ -20,6 +20,7 @@ class AccountEndpointTest extends EndpointTestCase
         self::assertDefaultEndpointPath(self::endpoint()->users(), '/api/account/users');
         self::assertDefaultEndpointPath(self::endpoint()->certificates(), '/api/account/certificates');
         self::assertDefaultEndpointPath(self::endpoint()->brandings(), '/api/account/brandings');
+        self::assertDefaultEndpointPath(self::endpoint()->smsSenders(), '/api/account/sms-senders');
         self::assertDefaultEndpointPath(self::endpoint()->messaging(), '/api/account/messaging');
         self::assertDefaultEndpointPath(self::endpoint()->signatureScenarios(), '/api/account/signature-scenarios');
         self::assertDefaultEndpointPath(self::endpoint()->identifyScenarios(), '/api/account/identify-scenarios');